move test_harness to mod
